collab: Extract tests into integration tests crate
