+++ crates/collab/src/db/queries/contributors.rs
@@ -2,7 +2,8 @@ use super::*;
 
 impl Database {
     /// Records that a given user has signed the CLA.
-    #[cfg(test)]
+    #[cfg(feature = "test-support")]
+
     pub async fn add_contributor(
         &self,
         github_login: &str,
