+++ crates/collab/tests/integration/notification_tests.rs
@@ -6,7 +6,7 @@ use parking_lot::Mutex;
 use pretty_assertions::assert_eq;
 use rpc::{Notification, proto};
 
-use crate::tests::TestServer;
+use crate::TestServer;
 
 #[gpui::test]
 async fn test_notifications(
