+++ crates/collab/tests/integration/random_project_collaboration_tests.rs
@@ -1,8 +1,8 @@
-use super::{RandomizedTest, TestClient, TestError, TestServer, UserTestPlan};
-use crate::{db::UserId, tests::run_randomized_test};
+use crate::{RandomizedTest, TestClient, TestError, TestServer, UserTestPlan, run_randomized_test};
 use anyhow::{Context as _, Result};
 use async_trait::async_trait;
 use call::ActiveCall;
+use collab::db::UserId;
 use collections::{BTreeMap, HashMap};
 use editor::Bias;
 use fs::{FakeFs, Fs as _};
@@ -33,10 +33,7 @@ use util::{
     rel_path::{RelPath, RelPathBuf, rel_path},
 };
 
-#[gpui::test(
-    iterations = 100,
-    on_failure = "crate::tests::save_randomized_test_plan"
-)]
+#[gpui::test(iterations = 100, on_failure = "crate::save_randomized_test_plan")]
 async fn test_random_project_collaboration(
     cx: &mut TestAppContext,
     executor: BackgroundExecutor,
